start mmio, generic putchar detection

+++ src/BIOSROM.hpp
@@ -15,12 +15,12 @@ public:
 
     BIOSROM() : MemBlockDevice(SIZE) {};
 
-    void write8([[maybe_unused]] u32 addr, [[maybe_unused]] u8 val) override
+    void write8([[maybe_unused]] u32 addr, [[maybe_unused]] u32 val) override
     {
         throw Panic("Trying to write to BIOS ROM");
     }
 
-    void write16([[maybe_unused]] u32 addr, [[maybe_unused]] u16 val) override
+    void write16([[maybe_unused]] u32 addr, [[maybe_unused]] u32 val) override
     {
         throw Panic("Trying to write to BIOS ROM");
     }
