+++ src/MemBlockDevice.hpp
@@ -16,8 +16,8 @@ public:
     u16 read16(u32 addr) override;
     u32 read32(u32 addr) override;
 
-    void write8(u32 addr, u8 val) override;
-    void write16(u32 addr, u16 val) override;
+    void write8(u32 addr, u32 val) override;
+    void write16(u32 addr, u32 val) override;
     void write32(u32 addr, u32 val) override;
 
     usize get_size();
